+++ libs/shared/models/src/index.ts
@@ -2,3 +2,4 @@ export { Filter } from './Filter';
 export { MovieFilter } from './MovieFilter';
 export type { NameEntityModel } from './NameEntityModel';
 export type { SelectableModel } from './SelectableModel';
+export { FilterType, FilterSettingsRaw, FilterSettings } from './FilterSettingsModel';
